+++ data/src/main/java/com/susu/data/remote/api/LedgerService.kt
@@ -26,7 +26,7 @@ interface LedgerService {
 
     @PATCH("ledgers/{id}")
     suspend fun editLedger(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Body ledgerRequest: LedgerRequest,
     ): ApiResult<LedgerResponse>
 
@@ -36,5 +36,5 @@ interface LedgerService {
     ): ApiResult<LedgerResponse>
 
     @DELETE("ledgers")
-    suspend fun deleteLedgerList(@Query("ids") idList: List<Int>): ApiResult<Unit>
+    suspend fun deleteLedgerList(@Query("ids") idList: List<Long>): ApiResult<Unit>
 }
